key loader: did: jwk: Ditch multibase did keys
