+++ environment.yml
@@ -40,5 +40,3 @@ dependencies:
     - PyJWT==2.8.0
     - werkzeug==2.2.2
     - cwt==2.7.1
-    - py-multibase==1.0.3
-    - py-multicodec==0.2.1
