+++ scitt_emulator/create_statement.py
@@ -1,5 +1,6 @@
 # Copyright (c) SCITT Authors
 # Licensed under the MIT License.
+import base64
 import pathlib
 import argparse
 from typing import Union, Optional
@@ -12,17 +13,10 @@ import pycose.keys.ec2
 from cryptography.hazmat.primitives.serialization import Encoding, PublicFormat
 from cryptography.hazmat.primitives.serialization import load_pem_private_key
 
-# NOTE These are unmaintained but the
-# https://github.com/hashberg-io/multiformats stuff and base58 modules don't
-# produce the same results:
-# https://grotto-networking.com/blog/posts/DID_Key.html#bug-in-multibase-library
-import multibase
-import multicodec
-
 # TODO jwcrypto is LGPLv3, is there another option with a permissive licence?
 import jwcrypto.jwk
 
-from scitt_emulator.did_helpers import DID_KEY_METHOD, MULTICODEC_HEX_P384_PUBLIC_KEY
+from scitt_emulator.did_helpers import DID_JWK_METHOD
 
 
 @pycose.headers.CoseHeaderAttribute.register_attribute()
@@ -102,22 +96,9 @@ def create_claim(
     cwt_cose_key_to_cose_key = cwt_cose_key.to_dict()
     sign1_message_key = pycose.keys.ec2.EC2Key.from_dict(cwt_cose_key_to_cose_key)
 
-    # If issuer was not given used did:key of public key
+    # If issuer was not given used did:jwk of public key
     if issuer is None:
-        multicodec_prefix_p_384 = "p384-pub"
-        multicodec.constants.NAME_TABLE[multicodec_prefix_p_384] = MULTICODEC_HEX_P384_PUBLIC_KEY
-        issuer = (
-            DID_KEY_METHOD
-            + multibase.encode(
-                "base58btc",
-                multicodec.add_prefix(
-                    multicodec_prefix_p_384,
-                    load_pem_private_key(key_as_pem_bytes, password=None)
-                    .public_key()
-                    .public_bytes(Encoding.X962, PublicFormat.CompressedPoint),
-                ),
-            ).decode()
-        )
+        issuer = DID_JWK_METHOD + base64.urlsafe_b64encode(key.export_public().encode()).decode()
 
     # CWT_Claims (label: 14 pending [CWT_CLAIM_COSE]): A CWT representing
     # the Issuer (iss) making the statement, and the Subject (sub) to
